Emonitor/requirements.txt
matplotlib
pandas
numpy
torch
torchvision
Pillow
opencv-python